requirements.txt
django
djangorestframework
djangorestframework-simplejwt 
celery
django-celery-beat 
pillow 
boto3
pandas